Delete confirmation dialogs added

+++ app/src/main/java/com/example/teachersapp/activity/StudentListActivity.java
@@ -4,6 +4,7 @@ import android.arch.lifecycle.ViewModelProviders;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.StaggeredGridLayoutManager;
@@ -22,13 +23,13 @@ import java.util.Objects;
 
 public class StudentListActivity extends AppCompatActivity {
     public static final int ADD_STUDENT_REQUEST = 1;
-    private static final String STATE_KEY = "students_activity";
+    public static final int DELETE_STUDENT_REQUEST = 2;
     private static final String TAG = "StudentListActivity";
 
     private StudentViewModel studentViewModel;
 
-    private RecyclerView recyclerView;
     private StudentAdapter adapter;
+    private int selectedContactPosition;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -36,22 +37,21 @@ public class StudentListActivity extends AppCompatActivity {
         setContentView(R.layout.activity_student_list);
         Objects.requireNonNull(getSupportActionBar()).setTitle(R.string.students);
 
-        recyclerView = findViewById(R.id.student_list);
+        RecyclerView recyclerView = findViewById(R.id.student_list);
         recyclerView.setHasFixedSize(true);
         recyclerView.setLayoutManager(new StaggeredGridLayoutManager(3, StaggeredGridLayoutManager.VERTICAL));
 
         adapter = new StudentAdapter();
-        adapter.setOnStudentClickListener(student -> {
+        adapter.setOnStudentClickListener((student, position) -> {
+            selectedContactPosition = position;
             Intent i = new Intent(StudentListActivity.this, StudentActivity.class);
             i.putExtra(Student.class.getCanonicalName(), student);
-            startActivity(i);
+            startActivityForResult(i, DELETE_STUDENT_REQUEST);
         });
 
         recyclerView.setAdapter(adapter);
         studentViewModel = ViewModelProviders.of(this).get(StudentViewModel.class);
-        studentViewModel.getAllStudents().observe(this, students -> {
-            adapter.setStudents(students);
-        });
+        studentViewModel.getAllStudents().observe(this, students -> adapter.setStudents(students));
 
         findViewById(R.id.add_student_fab).setOnClickListener(v -> {
             Log.d(TAG, "FAB: pressed!");
@@ -71,6 +71,10 @@ public class StudentListActivity extends AppCompatActivity {
             studentViewModel.insert(student);
             Toast.makeText(this, R.string.student_saved, Toast.LENGTH_SHORT).show();
         }
+        if (requestCode == DELETE_STUDENT_REQUEST && resultCode == RESULT_OK) {
+            studentViewModel.delete(adapter.getContactOnPosition(selectedContactPosition));
+            Toast.makeText(this, R.string.student_deleted, Toast.LENGTH_SHORT).show();
+        }
     }
 
     @Override
@@ -84,9 +88,24 @@ public class StudentListActivity extends AppCompatActivity {
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()) {
             case R.id.action_delete_all_students:
-                studentViewModel.deleteAllStudents();
+                deleteAllStudents();
+                return true;
+            default:
+                return super.onOptionsItemSelected(item);
         }
-        return super.onOptionsItemSelected(item);
+    }
+
+    private void deleteAllStudents() {
+        AlertDialog confirmationDialog = new AlertDialog.Builder(StudentListActivity.this)
+                .setMessage(R.string.delete_all_students_q)
+                .setPositiveButton(R.string.delete, (dialog, which) -> {
+                    studentViewModel.deleteAllStudents();
+                    Toast.makeText(StudentListActivity.this, R.string.all_students_deleted, Toast.LENGTH_SHORT).show();
+                    dialog.dismiss();
+                })
+                .setNegativeButton(R.string.cancel, ((dialog, which) -> dialog.dismiss()))
+                .create();
+        confirmationDialog.show();
     }
